Split BTree Dedupe out into its own file
using string() instead of sprintf() for getting the beast raw string
add a small perf win for forgetfulsyncmap for just a string key
